Solution
All submissions for this problem are available.NIT Trichy is a well-known institute, with lots of talented programmers, who are planning to make some teams amongst themselves and are searching for one more talented programmer. In search of a programmer, they provide you the opportunity by giving a following problem statement to solve.

Initially, every programmer is alone on his/her own team. When a programmer of one team sends an invitation to a programmer of another team, their teams combine together into a single team. If they already belong to the same team, then nothing changes.

These teams sometimes get too competitive. We say that two distinct terms T1 and T2 are competing if |size (T2) – size (T1)| >= c, where size (T) denotes the number of members in a team T and c denotes the competitive constant.

The code should accept and process some queries. The query follows the syntax as shown below:

1xy. Programmer x invites Programmer y to join their team.
2c. Find the number of pairs of teams which are currently competing, using the competitive constant c.
For example, given 3 programmers, and a query 121,

Solve the above problem and build a code which takes in two integers n and q and prints the answer for each query of type 2 in a separate line. You need to take the queries from the standard input as described in the input format section.